+++ changelog.md
@@ -5,6 +5,8 @@
   + `workspace.supportScheme`: `["file", "untitled", "git"]`
   + `diagnostics.disableScheme`: `["git"]`
 * `NEW` folding: support folding `---@alias`
+* `CHG` if `rootUri` or `workspaceFolder` is set to `ROOT` or `HOME`, this extension will refuse to load these directories and show an error message.
+* `CHG` show warning message when scanning more than 100,000 files.
 * `FIX` hover: can not union `table` with other basic types
 * `FIX` [#1125](https://github.com/sumneko/lua-language-server/issues/1125)
 * `FIX` [#1131](https://github.com/sumneko/lua-language-server/issues/1131)